% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_package.R
\docType{package}
\name{instantiate-package}
\alias{instantiate-package}
\title{instantiate: pre-compiled CmdStan models in R packages}
\description{
Similar to \href{https://mc-stan.org/rstantools/}{\code{rstantools}}
for \href{https://mc-stan.org/rstan/}{\code{rstan}},
the \code{instantiate} package builds pre-compiled
\href{https://mc-stan.org/users/interfaces/cmdstan}{\code{CmdStan}}
models into CRAN-ready statistical modeling R packages.
The models compile once during installation,
the executables live inside the file systems of their
respective packages, and users have the full power
and convenience of \href{https://mc-stan.org/cmdstanr/}{\code{CmdStanR}}
without any additional compilation after package installation.
This approach saves time and helps R package developers
migrate from \href{https://mc-stan.org/rstan/}{\code{rstan}}
to the more modern \href{https://mc-stan.org/cmdstanr/}{\code{CmdStanR}}.
}
\concept{help}
